Closer to being usable library
