+++ src/test_helpers/__init__.py
@@ -1,29 +1,48 @@
-# Copyright 2023 Canonical Ltd.
+# Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
 """Utils for observability tests."""
 
-from .tempo_helpers import *
+from .common import (
+    RelationData,
+    UnitRelationData,
+    deploy_and_configure_minio,
+    deploy_literal_bundle,
+    get_application_ip,
+    get_content,
+    get_databags,
+    get_relation_by_endpoint,
+    get_relation_data,
+    get_unit_address,
+    get_unit_info,
+    purge,
+    run_command,
+)
+from .tempo_helpers import (
+    deploy_tempo_cluster,
+    get_traces,
+    get_traces_patiently,
+    tempo_coordinator_charm_and_channel,
+    tempo_worker_charm_and_channel,
+)
 
 __all__ = [
-    'purge',
-    'get_unit_info',
-    'get_relation_by_endpoint',
-    'get_content',
-    'get_databags',
-    'get_relation_data',
-    'deploy_literal_bundle',
-    'run_command',
-    'present_facade',
-    'get_unit_address',
-    'deploy_and_configure_minio',
-    'tempo_coordinator_charm_and_channel',
-    'tempo_worker_charm_and_channel',
-    'deploy_cluster',
-    'get_traces',
-    'get_traces_patiently',
-    'get_application_ip',
-    'emit_trace',
-    'UnitRelationData',
-    'RelationData',
+    "purge",
+    "get_unit_info",
+    "get_relation_by_endpoint",
+    "get_content",
+    "get_databags",
+    "get_relation_data",
+    "deploy_literal_bundle",
+    "run_command",
+    "get_unit_address",
+    "deploy_and_configure_minio",
+    "tempo_coordinator_charm_and_channel",
+    "tempo_worker_charm_and_channel",
+    "deploy_tempo_cluster",
+    "get_traces",
+    "get_traces_patiently",
+    "get_application_ip",
+    "UnitRelationData",
+    "RelationData",
 ]
