+++ tox.ini
@@ -2,7 +2,7 @@
 # See LICENSE file for licensing details.
 
 [tox]
-envlist = lint, static, unit
+envlist = lint, unit
 isolated_build=true
 
 [vars]
@@ -41,18 +41,6 @@ commands =
     ruff check {[vars]all_path}
     black --check --diff {[vars]all_path}
 
-[testenv:static]
-description = Run static analysis checks
-deps =
-    ops[testing]
-    PyYAML
-    typing_extensions
-    pyright
-setenv =
-    PYTHONPATH = {toxinidir}:{toxinidir}/lib
-commands =
-    pyright {[vars]src_path}
-
 [testenv:unit]
 description = Run unit tests
 deps =
@@ -67,6 +55,8 @@ deps =
     jsonschema
     lightkube>=v0.15.4
     lightkube-models==1.24.1.4
+    juju
+    pytest_operator
 allowlist_externals =
     /usr/bin/env
     charmcraft
